requirements.txt
asgiref==3.7.2
channels==4.0.0
Django==4.2.6
djangorestframework==3.14.0
djangorestframework_simplejwt==5.3.1
firebase_admin==6.5.0
nltk==3.8.1
python-dotenv==1.0.1
gunicorn==21.2.0
daphne==3.0.2
django-cors-headers==4.3.1
